wifi: Expect P2P API failure for unknown peers
The ReinvokePersistentGroup and ProvisionDiscoveryWithParams tests
were using a fake MAC address that does not exist in the supplicant's
peer list. This causes the underlying API calls to fail, as expected.

This commit updates the tests to correctly assert for a failure status
in this scenario, aligning them with the expected supplicant behavior.

Bug: 425610011
Test: atest SupplicantP2pIfaceAidlTest

CRs-Fixed: 4151993
Change-Id: I156d13ee603a900d27bf97d271d497ae56130bc5
Signed-off-by: Utkarsh Soni <quic_usoni@quicinc.com>
(cherry picked from commit 9c7477fadf300ba4ace602d295b14f185d551278)